Refactor packages content
