+++ icedata-v2/src/main/java/biz/icecat/icedatav2/models/entity/SupplierEntity.java
@@ -1,4 +1,4 @@
-package biz.icecat.icedatav2.persistence.entity;
+package biz.icecat.icedatav2.models.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
