--- icedata-v2/src/main/java/biz/icecat/icedatav2/persistence/repository/SupplierRepository.java
@@ -1,8 +1,0 @@
-package biz.icecat.icedatav2.persistence.repository;
-
-import biz.icecat.icedatav2.persistence.entity.SupplierEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SupplierRepository extends JpaRepository<SupplierEntity, Long> {}
