+++ icedata-v2/src/main/java/biz/icecat/icedatav2/repository/SupplierRepository.java
@@ -1,0 +1,9 @@
+package biz.icecat.icedatav2.repository;
+
+import biz.icecat.icedatav2.models.entity.SupplierEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SupplierRepository extends JpaRepository<SupplierEntity, Long> {
+}
