+++ icedata-v2/src/main/java/biz/icecat/icedatav2/utils/FieldUtils.java
@@ -1,6 +1,6 @@
 package biz.icecat.icedatav2.utils;
 
-import biz.icecat.icedatav2.persistence.entity.SupplierEntity;
+import biz.icecat.icedatav2.models.entity.SupplierEntity;
 import biz.icecat.icedatav2.service.extractors.XmlAttributeBiConsumer;
 import lombok.experimental.UtilityClass;
